bridge code search refresh to target local source
