+++ crates/axon-services/src/query.rs
@@ -1,70 +1,27 @@
-use crate::context::ServiceContext;
-use crate::document::{
-    decode_document_cursor_backend, is_stale, paginate_document, read_latest_stored_source,
-};
 use crate::events::{LogLevel, ServiceEvent, emit, synthesis_delta_handler_infallible};
-use crate::scrape as scrape_svc;
 use crate::types::{
-    AskResult, CodeSearchCaller, CodeSearchFreshness, CodeSearchOptions, CodeSearchResult,
-    DocumentBackend, EvaluateResult, Pagination, QueryHit, QueryResult, RetrieveOptions,
-    RetrieveResult, ServiceRetrieveVariantError, SuggestResult, Suggestion,
+    AskResult, DocumentBackend, EvaluateResult, Pagination, QueryHit, QueryResult, RetrieveResult,
+    ServiceRetrieveVariantError, SuggestResult, Suggestion,
 };
-use axon_code_index::config::validate_path_prefix;
-use axon_code_index::ensure::{EnsureFreshOptions, ensure_fresh_with_progress};
-use axon_code_index::store::CodeIndexStore;
-use axon_code_index::{
-    CodeIndexIdentity, CodeSearchAllowedRoots, FreshnessWarning, ReindexProgressSink,
-};
-use axon_core::config::{Config, ConfigOverrides, ScrapeFormat};
+use axon_core::config::Config;
 use axon_core::error::{ServiceError, diagnostics_from_error};
 use axon_vector::ops::commands::ask::{ask_result, ask_result_with_deltas};
 use axon_vector::ops::commands::discover_crawl_suggestions;
 use axon_vector::ops::commands::evaluate_result;
 use axon_vector::ops::commands::query_hits;
-use axon_vector::ops::commands::{CodeSearchVectorRequest, code_search_hits};
-use axon_vector::ops::qdrant::{DirectRetrieveResult, retrieve_result};
+use axon_vector::ops::qdrant::DirectRetrieveResult;
 use std::error::Error;
-use std::path::{Path, PathBuf};
-use std::process::Command;
-use std::time::Duration;
 use tokio::sync::mpsc;
 
-const RETRIEVE_STALE_AFTER: Duration = Duration::from_secs(24 * 60 * 60);
-const MAX_CODE_SEARCH_QUERY_LEN_BYTES: usize = 64 * 1024;
-const CODE_SEARCH_GIT_TIMEOUT: Duration = Duration::from_secs(5);
-const CODE_SEARCH_REFRESH_BACKEND_ENV: &str = "AXON_CODE_SEARCH_REFRESH_BACKEND";
+pub use self::code_search::{
+    CodeSearchProjectResult, CodeSearchRefreshBackend, CodeSearchRefreshResult, code_search,
+    code_search_with_progress, refresh_code_search_index, refresh_code_search_index_with_backend,
+    refresh_code_search_index_with_progress, resolve_code_search_project,
+};
+pub use self::retrieve::retrieve;
 
-struct ResolvedDocument {
-    backend: DocumentBackend,
-    content: String,
-    chunk_count: usize,
-    matched_url: Option<String>,
-    warnings: Vec<String>,
-    variant_errors: Vec<ServiceRetrieveVariantError>,
-    source_truncated: bool,
-    refresh_status: Option<String>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CodeSearchRefreshBackend {
-    LegacyCodeIndex,
-    TargetLocalSource,
-}
-
-impl CodeSearchRefreshBackend {
-    fn from_config_value(value: Option<&str>) -> Result<Self, String> {
-        let Some(value) = value.map(str::trim).filter(|value| !value.is_empty()) else {
-            return Ok(Self::LegacyCodeIndex);
-        };
-        match value {
-            "legacy" | "legacy-code-index" | "code-index" => Ok(Self::LegacyCodeIndex),
-            "target-local" | "target-local-source" | "source-local" => Ok(Self::TargetLocalSource),
-            _ => Err(format!(
-                "unsupported {CODE_SEARCH_REFRESH_BACKEND_ENV} value `{value}`"
-            )),
-        }
-    }
-}
+mod code_search;
+mod retrieve;
 
 fn wrap_service_error(
     message: String,
@@ -195,622 +152,6 @@ pub async fn query(
         })?;
     Ok(QueryResult { results })
 }
-
-/// Search one local git checkout after optionally refreshing its local-code vectors.
-#[must_use = "code_search returns a Result that should be handled"]
-pub async fn code_search(
-    ctx: &ServiceContext,
-    text: &str,
-    opts: CodeSearchOptions,
-) -> Result<CodeSearchResult, Box<dyn Error + Send + Sync>> {
-    code_search_with_progress(ctx, text, opts, None).await
-}
-
-#[must_use = "code_search_with_progress returns a Result that should be handled"]
-pub async fn code_search_with_progress(
-    ctx: &ServiceContext,
-    text: &str,
-    opts: CodeSearchOptions,
-    progress: Option<&dyn ReindexProgressSink>,
-) -> Result<CodeSearchResult, Box<dyn Error + Send + Sync>> {
-    if text.len() > MAX_CODE_SEARCH_QUERY_LEN_BYTES {
-        return Err(format!(
-            "code_search query exceeds {MAX_CODE_SEARCH_QUERY_LEN_BYTES}-byte cap (got {} bytes)",
-            text.len()
-        )
-        .into());
-    }
-
-    let path_prefix = opts
-        .path_prefix
-        .as_deref()
-        .map(validate_path_prefix)
-        .transpose()?
-        .flatten();
-    let root = resolve_code_search_root(opts.cwd.as_deref(), opts.caller).await?;
-    let identity = code_search_identity(ctx.cfg(), root).await;
-    let freshness =
-        resolve_code_search_freshness_with_progress(ctx, &identity, opts.ensure_fresh, progress)
-            .await;
-    let Some(committed_generation) = code_search_committed_generation(ctx, &identity).await? else {
-        return Ok(code_search_missing_index_result(text, freshness));
-    };
-
-    let results = code_search_hits(
-        ctx.cfg(),
-        CodeSearchVectorRequest {
-            query: text,
-            limit: opts.limit.max(1),
-            offset: opts.offset,
-            project_key: &identity.project_key,
-            generation: committed_generation,
-            path_prefix: path_prefix.as_deref(),
-        },
-    )
-    .await
-    .map_err(|e| -> Box<dyn Error + Send + Sync> {
-        let message = format!(
-            "code_search vector query failed for {}: {e}",
-            text.chars().take(80).collect::<String>()
-        );
-        wrap_service_error(message, e.as_ref())
-    })?;
-
-    Ok(CodeSearchResult {
-        query: text.to_string(),
-        content_trust: "untrusted_local_code".to_string(),
-        results,
-        freshness,
-    })
-}
-
-/// Extract the SQLite pool backing the code index from the service runtime.
-/// Code-index functions take the raw pool (not `ServiceContext`) so they live
-/// below the services layer without a dependency cycle.
-fn code_index_pool(ctx: &ServiceContext) -> Result<sqlx::SqlitePool, Box<dyn Error + Send + Sync>> {
-    ctx.jobs
-        .sqlite_pool()
-        .map(|pool| pool.as_ref().clone())
-        .ok_or_else(|| "code search requires a SQLite service runtime".into())
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize)]
-pub struct CodeSearchRefreshResult {
-    pub project_root: PathBuf,
-    pub project_key: String,
-    pub generation: Option<i64>,
-    pub freshness: CodeSearchFreshness,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize)]
-pub struct CodeSearchProjectResult {
-    pub project_root: PathBuf,
-    pub project_key: String,
-}
-
-#[must_use = "resolve_code_search_project returns a Result that should be handled"]
-pub async fn resolve_code_search_project(
-    ctx: &ServiceContext,
-    cwd: Option<&Path>,
-    caller: CodeSearchCaller,
-) -> Result<CodeSearchProjectResult, Box<dyn Error + Send + Sync>> {
-    let root = resolve_code_search_root(cwd, caller).await?;
-    let identity = code_search_identity(ctx.cfg(), root).await;
-    Ok(CodeSearchProjectResult {
-        project_root: identity.project_root,
-        project_key: identity.project_key,
-    })
-}
-
-#[must_use = "refresh_code_search_index returns a Result that should be handled"]
-pub async fn refresh_code_search_index(
-    ctx: &ServiceContext,
-    cwd: Option<&Path>,
-    caller: CodeSearchCaller,
-) -> Result<CodeSearchRefreshResult, Box<dyn Error + Send + Sync>> {
-    refresh_code_search_index_with_progress(ctx, cwd, caller, None).await
-}
-
-#[must_use = "refresh_code_search_index_with_progress returns a Result that should be handled"]
-pub async fn refresh_code_search_index_with_progress(
-    ctx: &ServiceContext,
-    cwd: Option<&Path>,
-    caller: CodeSearchCaller,
-    progress: Option<&dyn ReindexProgressSink>,
-) -> Result<CodeSearchRefreshResult, Box<dyn Error + Send + Sync>> {
-    match selected_code_search_refresh_backend()? {
-        CodeSearchRefreshBackend::LegacyCodeIndex => {
-            refresh_legacy_code_search_index_with_progress(ctx, cwd, caller, progress).await
-        }
-        CodeSearchRefreshBackend::TargetLocalSource => {
-            refresh_target_local_code_search_index_with_progress(ctx, cwd, caller).await
-        }
-    }
-}
-
-fn selected_code_search_refresh_backend()
--> Result<CodeSearchRefreshBackend, Box<dyn Error + Send + Sync>> {
-    let value = std::env::var(CODE_SEARCH_REFRESH_BACKEND_ENV).ok();
-    Ok(CodeSearchRefreshBackend::from_config_value(
-        value.as_deref(),
-    )?)
-}
-
-async fn refresh_legacy_code_search_index_with_progress(
-    ctx: &ServiceContext,
-    cwd: Option<&Path>,
-    caller: CodeSearchCaller,
-    progress: Option<&dyn ReindexProgressSink>,
-) -> Result<CodeSearchRefreshResult, Box<dyn Error + Send + Sync>> {
-    let root = resolve_code_search_root(cwd, caller).await?;
-    let identity = code_search_identity(ctx.cfg(), root).await;
-    let freshness =
-        resolve_code_search_freshness_with_progress(ctx, &identity, true, progress).await;
-    let generation = code_search_committed_generation(ctx, &identity).await?;
-    Ok(CodeSearchRefreshResult {
-        project_root: identity.project_root.clone(),
-        project_key: identity.project_key.clone(),
-        generation,
-        freshness,
-    })
-}
-
-async fn refresh_target_local_code_search_index_with_progress(
-    ctx: &ServiceContext,
-    cwd: Option<&Path>,
-    caller: CodeSearchCaller,
-) -> Result<CodeSearchRefreshResult, Box<dyn Error + Send + Sync>> {
-    let root = resolve_code_search_root(cwd, caller).await?;
-    let identity = code_search_identity(ctx.cfg(), root).await;
-    Ok(CodeSearchRefreshResult {
-        project_root: identity.project_root,
-        project_key: identity.project_key,
-        generation: None,
-        freshness: code_search_freshness(
-            "stale",
-            Some(FreshnessWarning::Failed {
-                error: "target local source code-search refresh is not wired yet".to_string(),
-            }),
-            0,
-            0,
-        ),
-    })
-}
-
-async fn resolve_code_search_freshness_with_progress(
-    ctx: &ServiceContext,
-    identity: &CodeIndexIdentity,
-    ensure: bool,
-    progress: Option<&dyn ReindexProgressSink>,
-) -> CodeSearchFreshness {
-    if !ensure {
-        return code_search_freshness("skipped", None, 0, 0);
-    }
-
-    let pool = match code_index_pool(ctx) {
-        Ok(pool) => pool,
-        Err(err) => {
-            return code_search_freshness(
-                "stale",
-                Some(FreshnessWarning::Failed {
-                    error: err.to_string(),
-                }),
-                0,
-                0,
-            );
-        }
-    };
-
-    match ensure_fresh_with_progress(
-        ctx.cfg(),
-        pool,
-        identity,
-        EnsureFreshOptions::default(),
-        progress,
-    )
-    .await
-    {
-        Ok(outcome) => code_search_freshness(
-            "fresh",
-            outcome.warning,
-            outcome.indexed_files,
-            outcome.removed_files,
-        ),
-        Err(err) => code_search_freshness(
-            "stale",
-            Some(FreshnessWarning::Failed {
-                error: err.to_string(),
-            }),
-            0,
-            0,
-        ),
-    }
-}
-
-async fn code_search_committed_generation(
-    ctx: &ServiceContext,
-    identity: &CodeIndexIdentity,
-) -> Result<Option<i64>, Box<dyn Error + Send + Sync>> {
-    let store = CodeIndexStore::open_for_pool(code_index_pool(ctx)?).await?;
-    let generation = store.committed_generation(identity).await?.unwrap_or(0);
-    Ok((generation > 0).then_some(generation))
-}
-
-fn code_search_missing_index_result(
-    text: &str,
-    freshness: CodeSearchFreshness,
-) -> CodeSearchResult {
-    CodeSearchResult {
-        query: text.to_string(),
-        content_trust: "untrusted_local_code".to_string(),
-        results: Vec::new(),
-        freshness: code_search_missing_index_freshness(freshness),
-    }
-}
-
-fn code_search_freshness(
-    status: &str,
-    warning: Option<FreshnessWarning>,
-    indexed_files: usize,
-    removed_files: usize,
-) -> CodeSearchFreshness {
-    let status = if warning.is_some() { "stale" } else { status };
-    CodeSearchFreshness {
-        status: status.to_string(),
-        warning: warning.map(|warning| warning.message()),
-        indexed_files,
-        removed_files,
-    }
-}
-
-fn code_search_missing_index_freshness(mut freshness: CodeSearchFreshness) -> CodeSearchFreshness {
-    if freshness.warning.is_none() {
-        freshness.status = "stale".to_string();
-        freshness.warning = Some(FreshnessWarning::MissingCommittedIndex.message());
-    }
-    freshness
-}
-
-async fn resolve_code_search_root(
-    cwd: Option<&Path>,
-    caller: CodeSearchCaller,
-) -> Result<PathBuf, Box<dyn Error + Send + Sync>> {
-    let cwd = match (caller, cwd) {
-        (CodeSearchCaller::Cli, Some(cwd)) => cwd.to_path_buf(),
-        (CodeSearchCaller::Cli, None) => std::env::current_dir()?,
-        (CodeSearchCaller::Mcp, Some(cwd)) => cwd.to_path_buf(),
-        (CodeSearchCaller::Mcp, None) => {
-            return Err("code_search MCP requests must provide cwd".into());
-        }
-    };
-    let canonical_cwd =
-        std::fs::canonicalize(&cwd).map_err(|_| "code_search cwd could not be resolved")?;
-    let git_root = git_toplevel(&canonical_cwd).await?;
-    reject_unsafe_code_root(&git_root)?;
-    if matches!(caller, CodeSearchCaller::Mcp) {
-        let allowed = CodeSearchAllowedRoots::from_env()?;
-        if !allowed.contains(&git_root) {
-            return Err(code_search_outside_allowed_roots_message().into());
-        }
-    }
-    Ok(git_root)
-}
-
-fn code_search_outside_allowed_roots_message() -> &'static str {
-    "code_search cwd is outside AXON_CODE_SEARCH_ALLOWED_ROOTS"
-}
-
-async fn git_toplevel(cwd: &Path) -> Result<PathBuf, Box<dyn Error + Send + Sync>> {
-    let cwd = cwd.to_path_buf();
-    let output = tokio::time::timeout(
-        CODE_SEARCH_GIT_TIMEOUT,
-        tokio::task::spawn_blocking(move || {
-            Command::new("git")
-                .arg("-C")
-                .arg(cwd)
-                .args(["rev-parse", "--show-toplevel"])
-                .output()
-        }),
-    )
-    .await
-    .map_err(|_| "git rev-parse timed out")?
-    .map_err(|e| format!("git rev-parse task failed: {e}"))?
-    .map_err(|_| "code_search cwd is not inside a git checkout")?;
-    if !output.status.success() {
-        return Err("code_search cwd is not inside a git checkout".into());
-    }
-    let root = String::from_utf8(output.stdout)
-        .map_err(|e| format!("git rev-parse output was not UTF-8: {e}"))?;
-    let root = root.trim();
-    if root.is_empty() {
-        return Err("git rev-parse returned an empty repository root".into());
-    }
-    std::fs::canonicalize(root).map_err(Into::into)
-}
-
-fn reject_unsafe_code_root(root: &Path) -> Result<(), Box<dyn Error + Send + Sync>> {
-    if root == Path::new("/") {
-        return Err("code_search refuses to index filesystem root".into());
-    }
-    if let Some(home) = std::env::var_os("HOME").map(PathBuf::from)
-        && root == home.as_path()
-    {
-        return Err("code_search refuses to index HOME directly".into());
-    }
-    Ok(())
-}
-
-async fn code_search_identity(cfg: &Config, project_root: PathBuf) -> CodeIndexIdentity {
-    let origin = code_search_project_origin(&project_root).await;
-    let embedder = if cfg.tei_url.trim().is_empty() {
-        "tei".to_string()
-    } else {
-        cfg.tei_url.clone()
-    };
-    CodeIndexIdentity::new(project_root, origin, &cfg.collection, &embedder)
-}
-
-async fn code_search_project_origin(project_root: &Path) -> String {
-    let remote = match git_remote_origin(project_root).await {
-        Ok(Some(remote)) => remote,
-        Ok(None) => "git:no-origin".to_string(),
-        Err(error) => {
-            tracing::warn!(
-                %error,
-                project_root = %project_root.display(),
-                "code_search git remote origin lookup failed; using checkout-scoped fallback"
-            );
-            "git:no-origin".to_string()
-        }
-    };
-    // This seed is private input to the UUID project key. Only the derived key is
-    // stored in Qdrant payloads; the absolute root remains SQLite-only.
-    format!("{remote}\nworktree:{}", project_root.display())
-}
-
-async fn git_remote_origin(project_root: &Path) -> Result<Option<String>, String> {
-    let project_root = project_root.to_path_buf();
-    let output = tokio::time::timeout(
-        CODE_SEARCH_GIT_TIMEOUT,
-        tokio::task::spawn_blocking(move || {
-            Command::new("git")
-                .arg("-C")
-                .arg(project_root)
-                .args(["config", "--get", "remote.origin.url"])
-                .output()
-        }),
-    )
-    .await
-    .map_err(|_| "git remote origin lookup timed out".to_string())?
-    .map_err(|err| format!("git remote origin lookup task failed: {err}"))?
-    .map_err(|err| format!("git remote origin lookup failed to spawn git: {err}"))?;
-    if !output.status.success() {
-        return Ok(None);
-    }
-    let origin = String::from_utf8(output.stdout)
-        .map_err(|err| format!("git remote origin output was not UTF-8: {err}"))?;
-    let origin = origin.trim();
-    Ok((!origin.is_empty()).then(|| format!("git:{origin}")))
-}
-
-/// Retrieve stored document chunks for a URL.
-#[must_use = "retrieve returns a Result that should be handled"]
-pub async fn retrieve(
-    cfg: &Config,
-    url: &str,
-    opts: RetrieveOptions,
-) -> Result<RetrieveResult, Box<dyn Error + Send + Sync>> {
-    if url.starts_with("local-code://") {
-        return Err("local-code documents are only available through code_search".into());
-    }
-    let pinned_backend = decode_document_cursor_backend(opts.cursor.as_deref()).map_err(
-        |e| -> Box<dyn Error + Send + Sync> {
-            format!("invalid retrieve cursor for {url}: {e}").into()
-        },
-    )?;
-    let resolved = resolve_document(cfg, url, opts.max_points, pinned_backend).await?;
-    let page = paginate_document(
-        &resolved.content,
-        opts.cursor.as_deref(),
-        opts.token_budget,
-        resolved.backend,
-    )
-    .map_err(|e| -> Box<dyn Error + Send + Sync> {
-        format!("paginate retrieve result for {url}: {e}").into()
-    })?;
-    Ok(RetrieveResult {
-        chunk_count: resolved.chunk_count,
-        content: page.content,
-        requested_url: Some(url.to_string()),
-        matched_url: resolved.matched_url,
-        truncated: page.truncated || resolved.source_truncated,
-        warnings: resolved.warnings,
-        variant_errors: resolved.variant_errors,
-        token_estimate: page.token_estimate,
-        next_cursor: page.next_cursor,
-        remaining_tokens_estimate: page.remaining_tokens_estimate,
-        backend: Some(page.backend),
-        refresh_status: resolved.refresh_status,
-    })
-}
-
-async fn resolve_document(
-    cfg: &Config,
-    url: &str,
-    max_points: Option<usize>,
-    pinned_backend: Option<DocumentBackend>,
-) -> Result<ResolvedDocument, Box<dyn Error + Send + Sync>> {
-    if let Some(backend) = pinned_backend {
-        return match backend {
-            DocumentBackend::Qdrant => resolve_qdrant_document(cfg, url, max_points)
-                .await?
-                .ok_or_else(|| {
-                    "retrieve cursor requires qdrant backend but no stored chunks exist"
-                        .to_string()
-                        .into()
-                }),
-            DocumentBackend::StoredSource => resolve_stored_source_document(cfg, url)
-                .await?
-                .ok_or_else(|| {
-                    "retrieve cursor requires stored_source backend but no source file exists"
-                        .to_string()
-                        .into()
-                }),
-            DocumentBackend::LiveScrape => resolve_live_scrape_document(cfg, url, "cursor").await,
-        };
-    }
-
-    let mut qdrant_error: Option<String> = None;
-    match resolve_qdrant_document(cfg, url, max_points).await {
-        Ok(Some(qdrant)) => return Ok(qdrant),
-        Ok(None) => {}
-        Err(err) => qdrant_error = Some(err.to_string()),
-    }
-
-    if let Some(stored) = resolve_stored_source_document(cfg, url).await? {
-        if stored.refresh_status.as_deref() == Some("stale") {
-            match resolve_live_scrape_document(cfg, url, "stale").await {
-                Ok(mut refreshed) => {
-                    refreshed.warnings.extend(stored.warnings);
-                    if let Some(err) = qdrant_error {
-                        refreshed
-                            .warnings
-                            .push(format!("qdrant backend unavailable during retrieve: {err}"));
-                    }
-                    return Ok(refreshed);
-                }
-                Err(err) => {
-                    let mut stale = stored;
-                    stale.warnings.push(format!(
-                        "live scrape refresh failed; falling back to stale stored source: {err}"
-                    ));
-                    if let Some(qdrant_err) = qdrant_error {
-                        stale.warnings.push(format!(
-                            "qdrant backend unavailable during retrieve: {qdrant_err}"
-                        ));
-                    }
-                    return Ok(stale);
-                }
-            }
-        }
-        let mut stored = stored;
-        if let Some(err) = qdrant_error {
-            stored
-                .warnings
-                .push(format!("qdrant backend unavailable during retrieve: {err}"));
-        }
-        return Ok(stored);
-    }
-
-    let mut live = resolve_live_scrape_document(cfg, url, "miss").await?;
-    if let Some(err) = qdrant_error {
-        live.warnings
-            .push(format!("qdrant backend unavailable during retrieve: {err}"));
-    }
-    Ok(live)
-}
-
-async fn resolve_qdrant_document(
-    cfg: &Config,
-    url: &str,
-    max_points: Option<usize>,
-) -> Result<Option<ResolvedDocument>, Box<dyn Error + Send + Sync>> {
-    let result = retrieve_result(cfg, url, max_points).await.map_err(
-        |e| -> Box<dyn Error + Send + Sync> {
-            format!("qdrant retrieve failed for {url}: {e}").into()
-        },
-    )?;
-    if result.chunk_count == 0 {
-        return Ok(None);
-    }
-    let mapped = map_direct_retrieve_result(result);
-    Ok(Some(ResolvedDocument {
-        backend: DocumentBackend::Qdrant,
-        content: mapped.content,
-        chunk_count: mapped.chunk_count,
-        matched_url: mapped.matched_url,
-        warnings: mapped.warnings,
-        variant_errors: mapped.variant_errors,
-        source_truncated: mapped.truncated,
-        refresh_status: None,
-    }))
-}
-
-async fn resolve_stored_source_document(
-    cfg: &Config,
-    url: &str,
-) -> Result<Option<ResolvedDocument>, Box<dyn Error + Send + Sync>> {
-    let Some(stored) = read_latest_stored_source(&cfg.output_dir, url)
-        .await
-        .map_err(|e| -> Box<dyn Error + Send + Sync> {
-            format!("stored source lookup failed for {url}: {e}").into()
-        })?
-    else {
-        return Ok(None);
-    };
-    let stale = is_stale(stored.modified_at, RETRIEVE_STALE_AFTER);
-    let mut warnings = Vec::new();
-    if stale {
-        warnings.push(format!(
-            "stored source is stale (> {} hours old); attempting live refresh",
-            RETRIEVE_STALE_AFTER.as_secs() / 3600
-        ));
-    }
-    warnings.push(format!(
-        "using stored source file {}",
-        stored.path.display()
-    ));
-    Ok(Some(ResolvedDocument {
-        backend: DocumentBackend::StoredSource,
-        content: stored.content,
-        chunk_count: 0,
-        matched_url: Some(url.to_string()),
-        warnings,
-        variant_errors: Vec::new(),
-        source_truncated: false,
-        refresh_status: stale.then(|| "stale".to_string()),
-    }))
-}
-
-async fn resolve_live_scrape_document(
-    cfg: &Config,
-    url: &str,
-    reason: &str,
-) -> Result<ResolvedDocument, Box<dyn Error + Send + Sync>> {
-    let scrape_cfg = cfg.apply_overrides(&ConfigOverrides {
-        format: Some(ScrapeFormat::Markdown),
-        output_path: Some(None),
-        ..ConfigOverrides::default()
-    });
-    let result = scrape_svc::scrape(&scrape_cfg, url, None).await.map_err(
-        |e| -> Box<dyn Error + Send + Sync> {
-            format!("live scrape refresh failed for {url}: {e}").into()
-        },
-    )?;
-    let refresh_status = match reason {
-        "stale" => Some("refreshed_stale".to_string()),
-        "miss" => Some("refreshed_missing".to_string()),
-        "cursor" => Some("cursor_live_scrape".to_string()),
-        _ => Some(reason.to_string()),
-    };
-    let warning = match reason {
-        "stale" => "served fresh live scrape because stored source was stale",
-        "miss" => "served fresh live scrape because no indexed or stored content was available",
-        "cursor" => "continued retrieve via live scrape backend",
-        _ => "served fresh live scrape content",
-    };
-    Ok(ResolvedDocument {
-        backend: DocumentBackend::LiveScrape,
-        content: result.output,
-        chunk_count: 0,
-        matched_url: Some(result.url),
-        warnings: vec![warning.to_string()],
-        variant_errors: Vec::new(),
-        source_truncated: false,
-        refresh_status,
-    })
-}
-
 /// RAG ask: retrieve relevant context, then answer with LLM.
 ///
 /// When `cfg.ask_stream` is true and `tx` is `Some`, synthesis tokens are
@@ -1055,6 +396,12 @@ fn is_well_formed_suggest_url(url: &str) -> bool {
     host.contains('.') && !host.starts_with('.') && !host.ends_with('.')
 }
 
+#[cfg(test)]
+#[path = "query/payload_tests.rs"]
+mod payload_tests;
+#[cfg(test)]
+#[path = "query/retrieve_tests.rs"]
+mod retrieve_tests;
 #[cfg(test)]
 #[path = "query_tests.rs"]
 mod tests;
