+++ crates/axon-services/src/test_support.rs
@@ -1,0 +1,86 @@
+use std::collections::HashMap;
+use std::error::Error;
+
+use async_trait::async_trait;
+use axon_jobs::backend::{BackendResult, JobKind, JobPayload};
+use axon_jobs::status::JobStatus;
+
+use crate::runtime::ServiceJobRuntime;
+
+#[derive(Default)]
+pub(crate) struct NoopServiceRuntime;
+
+#[async_trait]
+impl ServiceJobRuntime for NoopServiceRuntime {
+    fn mode_name(&self) -> &'static str {
+        "test"
+    }
+
+    async fn enqueue(&self, _payload: JobPayload) -> BackendResult<uuid::Uuid> {
+        Ok(uuid::Uuid::new_v4())
+    }
+
+    async fn wait_for_job(&self, _id: uuid::Uuid, _kind: JobKind) -> BackendResult<String> {
+        Ok("completed".to_string())
+    }
+
+    async fn job_errors(&self, _id: uuid::Uuid, _kind: JobKind) -> BackendResult<Option<String>> {
+        Ok(None)
+    }
+
+    async fn has_active_jobs(&self, _kind: JobKind) -> BackendResult<bool> {
+        Ok(false)
+    }
+
+    async fn list_jobs(
+        &self,
+        _kind: JobKind,
+        _limit: i64,
+        _offset: i64,
+    ) -> Result<Vec<crate::types::ServiceJob>, Box<dyn Error + Send + Sync>> {
+        Ok(Vec::new())
+    }
+
+    async fn job_status(
+        &self,
+        _kind: JobKind,
+        _id: uuid::Uuid,
+    ) -> Result<Option<crate::types::ServiceJob>, Box<dyn Error + Send + Sync>> {
+        Ok(None)
+    }
+
+    async fn cancel_job(
+        &self,
+        _kind: JobKind,
+        _id: uuid::Uuid,
+    ) -> Result<bool, Box<dyn Error + Send + Sync>> {
+        Ok(false)
+    }
+
+    async fn cleanup_jobs(&self, _kind: JobKind) -> Result<u64, Box<dyn Error + Send + Sync>> {
+        Ok(0)
+    }
+
+    async fn clear_jobs(&self, _kind: JobKind) -> Result<u64, Box<dyn Error + Send + Sync>> {
+        Ok(0)
+    }
+
+    async fn recover_jobs(
+        &self,
+        _kind: JobKind,
+        _stale_threshold_ms: i64,
+    ) -> Result<u64, Box<dyn Error + Send + Sync>> {
+        Ok(0)
+    }
+
+    async fn count_jobs(&self, _kind: JobKind) -> Result<i64, Box<dyn Error + Send + Sync>> {
+        Ok(0)
+    }
+
+    async fn count_jobs_by_status(
+        &self,
+        _kind: JobKind,
+    ) -> Result<HashMap<JobStatus, i64>, Box<dyn Error + Send + Sync>> {
+        Ok(HashMap::new())
+    }
+}
