+++ docs/reference/sources/adapter-scopes.md
@@ -14,7 +14,7 @@ Generated by `cargo xtask schemas adapters`.
 | `gitea` | `git` | `repo` | `repo`, `branch`, `issue`, `pull_request`, `release` | false | true | none |
 | `github` | `git` | `repo` | `repo`, `branch`, `issue`, `pull_request`, `release` | false | true | none |
 | `gitlab` | `git` | `repo` | `repo`, `branch`, `issue`, `merge_request`, `release` | false | true | none |
-| `local` | `local` | `directory` | `directory`, `file`, `workspace` | true | true | none |
+| `local` | `local` | `directory` | `directory`, `file`, `workspace`, `repo`, `map` | true | true | none |
 | `mcp` | `mcp_tool` | `tool` | `tool` | false | true | none |
 | `npm` | `registry` | `package` | `package`, `version` | false | true | none |
 | `pypi` | `registry` | `package` | `package`, `version` | false | true | none |
@@ -28,5 +28,5 @@ Generated by `cargo xtask schemas adapters`.
 
 | Path | SHA-256 |
 |---|---|
-| `crates/axon-route/src/capability.rs` | `sha256:397100f9ad536758ff4cb9db030ff6b235f44d3f44bad04e930f2f13719cb120` |
+| `crates/axon-route/src/capability.rs` | `sha256:ded9abecd6713fb7663445ca66506f5b58a216ef3e627a2574dc098a9da565bb` |
 | `docs/pipeline-unification/sources/adapter-scopes.md` | `sha256:095b357f4ff7855e508a7c1fdbef857a2cac60d2f726fb860153a14f631ad832` |
