+++ docs/reference/sources/vector-payload.md
@@ -9,7 +9,12 @@ Generated by `cargo xtask schemas vector-payload`.
 | `payload_contract_version` | `string; const=2026-07-01` |
 | `collection` | `string; minLength=1` |
 | `source_family` | `string; enum=code|web|package|session|graph|memory; minLength=1` |
+| `source_kind` | `string; minLength=1` |
+| `source_adapter` | `string; minLength=1` |
+| `source_scope` | `string; minLength=1` |
 | `source_id` | `string; minLength=1` |
+| `source_item_key` | `string; minLength=1` |
+| `item_canonical_uri` | `string; minLength=1` |
 | `source_generation` | `string; minLength=1` |
 | `document_id` | `string; minLength=1` |
 | `chunk_id` | `string; minLength=1` |
@@ -44,7 +49,7 @@ Payload validation applies metadata and locator guardrails before vector writes.
 
 | Family | Fields |
 |---|---|
-| `code` | `code_language`, `code_symbol_name`, `code_symbol_kind`, `code_file_type` |
+| `code` | `code_language`, `code_symbol_name`, `code_symbol_kind`, `code_file_type`, `manifest` |
 | `web` | `web_title`, `web_domain`, `web_status_code`, `web_depth` |
 | `package` | `package_ecosystem`, `package_name`, `package_version` |
 | `session` | `session_id`, `session_turn_index`, `session_tool_name`, `session_skill_name` |
@@ -105,8 +110,8 @@ Vector payload docs are paired with the DTO definitions in `docs/reference/api/s
 | Path | SHA-256 |
 |---|---|
 | `crates/axon-api/src/source/vector.rs` | `sha256:97c767a15c9f88a3a7278ac32c6b7b1a4fce722223e9dd3c5f59079fee69b9ac` |
-| `crates/axon-vectors/src/payload.rs` | `sha256:a22a875e26722ba297f4a94cb1e0b0aa4c35713d78462b5a5d745930f0408854` |
-| `crates/axon-vectors/src/point.rs` | `sha256:698c1a043423685b53859cb7669c3424f3ea5c2a605c4b3d4b2a69ca4ad5b0bc` |
+| `crates/axon-vectors/src/payload.rs` | `sha256:28f9e6691093ce9336883b5f218411b2bf393daed7c4be69a97e73407235b930` |
+| `crates/axon-vectors/src/point.rs` | `sha256:f3c695929d6b5029c376417f761784fe8c4895c2a3fafa0429c2c1bc4f35448c` |
 | `docs/pipeline-unification/schemas/vector-payload-schema.md` | `sha256:ef84517d66acc0e35fe0b432df3cee52fc07e6643006003eca0ec0c29a5cde7f` |
 | `docs/pipeline-unification/sources/chunking-contract.md` | `sha256:c05b4d85b293af0200445e89adf99db1db55d3cf2e7d003fa38844efb682d8d8` |
 | `docs/pipeline-unification/sources/metadata-payload.md` | `sha256:debc62b438633191ee4b56a7597828d49752b322d4dffc3ca786a08cc2d3a0f2` |
